feat: updates

+++ custom/cmd_interface.go
@@ -14,9 +14,8 @@ import (
 func parseCli(argc C.int, argv **C.char) *C.char {
 	args := make([]string, argc)
 	for i := 0; i < int(argc); i++ {
-		// fmt.Println("parseCli", C.GoString(*argv))
 		args[i] = C.GoString(*argv)
-		argv = (**C.char)(unsafe.Pointer(uintptr(unsafe.Pointer(argv)) + uintptr(unsafe.Sizeof(*argv))))
+		argv = (**C.char)(unsafe.Pointer(uintptr(unsafe.Pointer(argv)) + unsafe.Sizeof(*argv)))
 	}
 	err := cmd.ParseCli(args[1:])
 	if err != nil {
